pyproject.toml
[tool.poetry]
name = "cocktailadvisorchat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.9"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
weaviate-client = "^4.11.0"
openai = "^1.65.1"
pandas = "^2.2.3"
pydantic-settings = "^2.8.1"
langchain = "^0.3.19"
langchain-openai = "^0.3.7"
langchain-community = "^0.3.18"
langchain-weaviate = "^0.0.4"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
